Add time range to side filter
